Docker/yahoo_llm_project/llm_service/requirements.txt
flask==2.3.3
redis==4.6.0
psycopg2-binary==2.9.7
bert-score==0.3.13
python-dotenv==1.0.0
requests==2.31.0
# Ollama support via HTTP API - no additional packages needed
# google-generativeai==0.3.2  # Commented out - using Ollama now
# transformers removed to speed up build - bert-score will install minimal dependencies